test: move nested-worker unref tests to their own file
worker_threads.test.ts has pre-existing debug-build timeouts (the 100 MiB
eval-source-leak test, deep environmentData inherit) unrelated to this
change. Put the new tests in a dedicated file so they run in isolation.
